Split out form controller
