+++ lib/first_days_web/controllers/confirmation_agreement_controller.ex
@@ -1,0 +1,7 @@
+defmodule FirstDaysWeb.ConfirmationAgreementController do
+  use FirstDaysWeb, :controller
+
+  def confirmation_agreement_show(conn, _params) do
+    render conn, "confirmation_agreement_show.html"
+  end
+end
