+++ lib/first_days_web/controllers/feedback_controller.ex
@@ -1,0 +1,7 @@
+defmodule FirstDaysWeb.FeedbackController do
+  use FirstDaysWeb, :controller
+
+  def feedback_show(conn, _params) do
+    render conn, "feedback_show.html"
+  end
+end
