+++ lib/first_days_web/templates/document_checklist/document_checklist_edit.html.eex
@@ -1,0 +1,3 @@
+<h2>Edit Documents</h2>
+
+<%= render "document_checklist_form.html", Map.merge(assigns, %{action: document_checklist_path(@conn, :document_checklist_update), method: :put}) %>
