+++ lib/first_days_web/templates/document_checklist/document_checklist_show.html.eex
@@ -24,6 +24,6 @@
 
 <p>Many thanks</p>
 
-<%= link "Edit", to: form_path(@conn, :document_checklist_edit), class: "btn btn-primary" %>
+<%= link "Edit", to: document_checklist_path(@conn, :document_checklist_edit), class: "btn btn-primary" %>
 <span></span>
 <%= link "Home", to: page_path(@conn, :first_days_index), class: "btn btn-primary" %>
