--- lib/first_days_web/templates/form/document_checklist_edit.html.eex
@@ -1,3 +1,0 @@
-<h2>Edit Documents</h2>
-
-<%= render "document_checklist_form.html", Map.merge(assigns, %{action: form_path(@conn, :document_checklist_update), method: :put}) %>
