--- lib/first_days_web/templates/form/role_description_edit.html.eex
@@ -1,3 +1,0 @@
-<h2>Edit Role Description Answers</h2>
-
-<%= render "role_description_form.html", Map.merge(assigns, %{action: form_path(@conn, :role_description_update), method: :put}) %>
