--- lib/first_days_web/templates/form/role_description_new.html.eex
@@ -1,3 +1,0 @@
-<h2>Role Description</h2>
-
-<%= render "role_description_form.html", Map.merge(assigns, %{action: form_path(@conn, :role_description_create), method: :post}) %>
