+++ lib/first_days_web/templates/page/first_days_index.html.eex
@@ -6,18 +6,18 @@
 
 <br/>
 
-<div><%= link "1. Answer questions to help us create a role description for the volunteer", to: form_path(@conn, :role_description_edit) %></div>
+<div><%= link "1. Answer questions to help us create a role description for the volunteer", to: role_description_path(@conn, :role_description_new) %></div>
 <br/>
-<div><%= link "2. Role Description template", to: form_path(@conn, :role_description_show) %></div>
+<div><%= link "2. Role Description template", to: role_description_path(@conn, :role_description_show) %></div>
 <br/>
-<div><%= link "3. Volunteer agreement", to: form_path(@conn, :confirmation_agreement_show) %></div>
+<div><%= link "3. Volunteer agreement", to: confirmation_agreement_path(@conn, :confirmation_agreement_show) %></div>
 <br/>
-<div><%= link "4. Answer questions to help us create a list of documents the volunteer needs to provide", to: form_path(@conn, :document_checklist_edit) %></div>
+<div><%= link "4. Answer questions to help us create a list of documents the volunteer needs to provide", to: document_checklist_path(@conn, :document_checklist_edit) %></div>
 <br/>
-<div><%= link "5. Document Checklist template", to: form_path(@conn, :document_checklist_show) %></div>
+<div><%= link "5. Document Checklist template", to: document_checklist_path(@conn, :document_checklist_show) %></div>
 <br/>
-<div><%= link "6. Answer questions to help us create an email informing the volunteer of everything they need to prepare for their first day", to: form_path(@conn, :preparation_edit) %></div>
+<div><%= link "6. Answer questions to help us create an email informing the volunteer of everything they need to prepare for their first day", to: preparation_path(@conn, :preparation_edit) %></div>
 <br/>
-<div><%= link "7. First day preparation email template", to: form_path(@conn, :preparation_show) %></div>
+<div><%= link "7. First day preparation email template", to: preparation_path(@conn, :preparation_show) %></div>
 <br/>
-<div><%= link "8. Volunteer feedback email template", to: form_path(@conn, :feedback_show) %></div>
+<div><%= link "8. Volunteer feedback email template", to: feedback_path(@conn, :feedback_show) %></div>
