+++ lib/first_days_web/templates/preparation/preparation_edit.html.eex
@@ -1,0 +1,3 @@
+<h2>Edit Preparation Answers</h2>
+
+<%= render "preparation_form.html", Map.merge(assigns, %{action: preparation_path(@conn, :preparation_update), method: :put}) %>
