+++ lib/first_days_web/templates/preparation/preparation_new.html.eex
@@ -1,4 +1,4 @@
 <h2>Preparation</h2>
 <p>It's important that your volunteer is prepared for their first day and has completed or sent through any paperwork they need to fill out for the role.</p>
 
-<%= render "preparation_form.html", Map.merge(assigns, %{action: form_path(@conn, :preparation_create), method: :post}) %>
+<%= render "preparation_form.html", Map.merge(assigns, %{action: preparation_path(@conn, :preparation_create), method: :post}) %>
