+++ lib/first_days_web/templates/preparation/preparation_show.html.eex
@@ -14,6 +14,6 @@
 <p>We look forward to seeing you then.</p>
 <p>Many thanks</p>
 
-<%= link "Edit", to: form_path(@conn, :preparation_edit), class: "btn btn-primary" %>
+<%= link "Edit", to: preparation_path(@conn, :preparation_edit), class: "btn btn-primary" %>
 <span></span>
 <%= link "Home", to: page_path(@conn, :first_days_index), class: "btn btn-primary" %>
