+++ lib/first_days_web/templates/role_description/role_description_new.html.eex
@@ -1,0 +1,3 @@
+<h2>Role Description</h2>
+
+<%= render "role_description_form.html", Map.merge(assigns, %{action: role_description_path(@conn, :role_description_create), method: :post}) %>
