+++ lib/first_days_web/views/confirmation_agreement_view.ex
@@ -1,0 +1,3 @@
+defmodule FirstDaysWeb.ConfirmationAgreementView do
+  use FirstDaysWeb, :view
+end
