+++ lib/first_days_web/views/feedback_view.ex
@@ -1,0 +1,3 @@
+defmodule FirstDaysWeb.FeedbackView do
+  use FirstDaysWeb, :view
+end
