+++ lib/first_days_web/views/role_description_view.ex
@@ -1,0 +1,13 @@
+defmodule FirstDaysWeb.RoleDescriptionView do
+  use FirstDaysWeb, :view
+
+  def finance_skills do
+    [
+    "Day to day bookkeeping. e.g. managing receipts, petty cash etc",
+    "Analysing/communicating management accounts e.g. to the trustees",
+    "Financial forecasting",
+    "Payroll",
+    "Pensions, Insurances, Investments"
+    ]
+  end
+end
